requirements.txt
google-genai
openai
requests
numpy
Pillow
